requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.
fabric
BeautifulSoup==3.2.0
GPolyEncode==0.1.1
python-dateutil

mysql-python
PIL
South
django-model-utils      # https://github.com/carljm/django-model-utils

ElementTree # for markdown
mimeparse
Markdown
django-tastypie
# slumber					# http://slumber.in/ | http://slumber.readthedocs.org/en/latest/index.html
html2text               # Convert HTML to Markdown-formatted text, https://github.com/aaronsw/html2text
django_extensions       # shell_plus
django-flatblocks

#specify commands in individual apps and then run them using the djboss command-line interface
django-boss

django-coffeescript     # https://github.com/andreyfedoseev/django-coffeescript
pystache
gunicorn                # python server, http://gunicorn.org/